Initial Commit
Task2 - Complete code to split the table and append data onto the new sheets as required by task. Requires optimisation of the code to accomodate for more items and data to be split by date rather than item name, as current algorithm looks for 'ITEM1' string each time and splits data according to that to append to a new sheet.