Changed variables to accomodate for larger dataset
specified date_range, which can be changed depending on data set. Likewise, specified num_rows of data.
Added a loop for adding more items to each set of data to be appended into a new sheet.
item_array and sheetNames are still fixed lists/arrays which will need to be amended to accomodate for more items and sheets to be created.